% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd.R
\name{spark_collect}
\alias{spark_collect}
\title{Collect elements of an RDD}
\usage{
spark_collect(x, flatten = TRUE)
}
\arguments{
\item{x}{The RDD to collect}

\item{flatten}{FALSE if the list should not flattened}

\item{...}{Other optional arguments to collect}
}
\value{
a list containing elements in the RDD
}
\description{
\code{collect} returns a list that contains all of the elements in this RDD.
}

