% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd.R
\name{spark_lapply}
\alias{spark_lapply}
\title{Apply a function to all elements}
\usage{
spark_lapply(X, FUN)
}
\arguments{
\item{X}{The RDD to apply the transformation.}

\item{FUN}{the transformation to apply on each element}
}
\value{
a new RDD created by the transformation.
}
\description{
This function creates a new RDD by applying the given transformation to all
elements of the given RDD
}

