% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_context.R
\name{spark_parallelize}
\alias{spark_parallelize}
\title{Create an RDD from a homogeneous list or vector.}
\usage{
spark_parallelize(sc, coll, numSlices = 1)
}
\arguments{
\item{sc}{SparkContext to use}

\item{coll}{collection to parallelize}

\item{numSlices}{number of partitions to create in the RDD}
}
\value{
an RDD created from this collection
}
\description{
This function creates an RDD from a local homogeneous list in R. The elements
in the list are split into \code{numSlices} slices and distributed to nodes
in the cluster.
}

